+++ src/main/kotlin/com/padovese/assessment/converter/RiskProfileDTOConverter.kt
@@ -6,7 +6,7 @@ import com.padovese.assessment.model.RiskProfile
 object RiskProfileDTOConverter {
 
     fun toRiskProfile(riskProfileDTO: RiskProfileDTO) : RiskProfile{
-        val baseScore = riskProfileDTO.riskQuestions.sum().toLong()
+        val baseScore = riskProfileDTO.riskQuestions.sum()
 
         return RiskProfile(
                 age = riskProfileDTO.age,
@@ -15,8 +15,8 @@ object RiskProfileDTOConverter {
                 dependents = riskProfileDTO.dependents,
                 maritalStatus = riskProfileDTO.maritalStatus,
                 vehicle = riskProfileDTO.vehicle,
-                vehicleBinary = if(riskProfileDTO.vehicle != null) 1L else 0L,
-                houseBinary = if(riskProfileDTO.house != null) 1L else 0L,
+                vehicleBinary = if(riskProfileDTO.vehicle != null) 1 else 0,
+                houseBinary = if(riskProfileDTO.house != null) 1 else 0,
                 autoScore = baseScore,
                 disabilityScore = baseScore,
                 homeScore = baseScore,
