+++ src/main/kotlin/com/padovese/assessment/dto/YearDTO.kt
@@ -1,5 +1,5 @@
 package com.padovese.assessment.dto
 
 data class YearDTO(
-        val year : Long
+        val year: Int
 )
