+++ src/main/kotlin/com/padovese/assessment/validator/Ineligible.kt
@@ -6,18 +6,18 @@ import org.springframework.stereotype.Component
 @Component
 class Ineligible {
     fun validate(riskProfile: RiskProfile) {
-        if(riskProfile.income == 0L)
-            riskProfile.disabilityScore = Long.MIN_VALUE
+        if(riskProfile.income == 0)
+            riskProfile.disabilityScore = Int.MIN_VALUE
 
-        if(riskProfile.houseBinary == 0L)
-            riskProfile.homeScore = Long.MIN_VALUE
+        if(riskProfile.houseBinary == 0)
+            riskProfile.homeScore = Int.MIN_VALUE
 
-        if(riskProfile.vehicleBinary == 0L)
-            riskProfile.autoScore = Long.MIN_VALUE
+        if(riskProfile.vehicleBinary == 0)
+            riskProfile.autoScore = Int.MIN_VALUE
 
         if(riskProfile.age > 60){
-            riskProfile.disabilityScore = Long.MIN_VALUE
-            riskProfile.lifeScore = Long.MIN_VALUE
+            riskProfile.disabilityScore = Int.MIN_VALUE
+            riskProfile.lifeScore = Int.MIN_VALUE
         }
     }
 }
